ADD: anchor tags in titles and images, Editor's pick
I had to erase all information in featured stories, top story, opinion
and more stories to set up all anchor tags in the images and titles.
They work fine now.

Also reuse article code to finish all articles in the "Editor's pick"
section. The title of the section is still missing the logo of the
company.